feat(rpc): Remove spans specific code (#83714)
- Removes spans specific code from the SearchResolver so that it can be
reused by logs
